Update documentation to avoid scheme-relative URLs
Per the HTML spec you can put a scheme-relative url, e.g. `//google.com` as the src of an iframe: https://html.spec.whatwg.org/multipage/iframe-embed-object.html#attr-iframe-src

But Zoid sniffs the protocol in isSameDomain (from cross-domain-utils) so it expects http or https. If a protocol is not provided, the preload page spinner spins infinitely as in #202.

This makes it explicit that a protocol should be provided.