Added logic on the vue driver to replace the prop name `styleObject` or `style-object` by a `style` prop.
In vue the prop name `style` is reserved thus if the user wants to pass in a style for the paypal button this won't work, instead the user now can pass in a `style-object` prop with the desire styling for the button.
To keep zoid generic, we are passing in the original prop name, so if zoid is used for other application the user will received the expected prop name in the component.